Move the logic of the conversation list to the left sidebar so all data is handled in one place
